+++ frontend/src/components/Attackers/InsertAttacker.vue
@@ -67,7 +67,7 @@
 </template>
 
 <script>
-import DropDown from '@/components/DropDown'
+import DropDown from '@/components/common/DropDown'
 import AttackerService from '@/services/attacker'
 export default {
   name: 'InsertAttacker',
