+++ frontend/src/components/Selections/SelectionRow.vue
