+++ frontend/src/components/common/TargetVillage.vue
